+++ CHANGELOG.md
@@ -36,6 +36,10 @@ This update accompanies the release of the [Telegram Bot API 6.6](https://core.t
  * Renamed `thumb` field in `TgBotLib\Objects\Telegram > Animation` to `thumbnail`
  * Renamed `thumb` field in `TgBotLib\Objects\Telegram > Audio` to `thumbnail`
  * Renamed `thumb` field in `TgBotLib\Objects\Telegram > Document` to `thumbnail`
+ * Renamed `thumb` field in `TgBotLib\Objects\Telegram > Sticker` to `thumbnail`
+
+## Fixed
+ * Fixed `TgBotLib\Objects\Telegram > Sticker > toArray()` to properly return the `premium_animation` field
 
 ## [6.5.2] - 2023-02-26